Relocate loadAssets into it's own file
